Move joinValuesWithLogicalCondition from ObjectUtil to StringUtil
