+++ server-spi/src/main/java/org/keycloak/utils/StringUtil.java
@@ -16,6 +16,8 @@
  */
 package org.keycloak.utils;
 
+import java.util.Collection;
+
 public class StringUtil {
 
     public static boolean isBlank(String str) {
@@ -24,6 +26,30 @@ public class StringUtil {
 
     public static boolean  isNotBlank(String str) {
         return str != null && !"".equals(str.trim());
+    }
+
+    /**
+     * Calling:
+     * <pre>joinValuesWithLogicalCondition("or", Arrays.asList("foo", "bar", "baz", "caz" ))</pre>
+     * will return "foo, bar, baz or caz"
+     *
+     * @param conditionText condition
+     * @param values values to be joined with the condition at the end
+     * @return see the example above
+     */
+    public static String joinValuesWithLogicalCondition(String conditionText, Collection<String> values) {
+        StringBuilder options = new StringBuilder();
+        int i = 1;
+        for (String o : values) {
+            if (i == values.size()) {
+                options.append(" " + conditionText + " ");
+            } else if (i > 1) {
+                options.append(", ");
+            }
+            options.append(o);
+            i++;
+        }
+        return options.toString();
     }
 
 }
